Refactor: ts마이그레이션(user레포지토리)
